Eliminate TIME_T_TYPE_DOUBLE ifdefs.
I held off getting rid of this in the initial pass because I thought it
might have some relevancy, but checking the original source I can't find
anything that sets it even in any of the Info-ZIP ports. Deleting this also
fixes the formatting that clang-format was stumbling over on this file.