Move langinfo.h include after unzip.h include.
It's non-intuitive why this is needed, but it's to do with the Large
File Support macros in unxcfg.h, which appear to change the size of
off_t (?) in the system headers. If we include langinfo.h first, the
type gets defined with the defaults and the program crashes because
internal type sizes aren't the same size between source files.

Really the LFS macros need to be defined on the compiler command line
(in CFLAGS) instead of in header files, but for the time being ...