Eliminate string concatenation in ExtractMsg.
In this format string we print the file type and optionally a newline at the
end of the string. Instead, let's do this as separate Info() calls.

Eliminating concatenation of the newline in particular means that Info() should
now be able to track when we reach an end-of-line simply by examining the last
character of the format string.