Get rid of last use of --More-- pause prompt.
This was previously used (I think) for the internal pager but there was one
last use of the function left and it's an obscure one. It looks like if
Info-ZIP encounters a .zip file containing an ASCII ^S character, it will pause
and wait for a keypress before proceeding. This only happens on extract. I
think this feature is sufficiently obscure that we can do without it.